requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 7.2.0
numpy == 1.19.0
opencv_python_headless == 4.2.0.34
pycocotools == 2.0.0
scipy == 1.5.1
torch == 1.6.0
torchvision == 0.7.0
tqdm == 4.48.0
